requirements.txt

fastapi
uvicorn[standard]
python-dotenv
pydantic
anthropic
deepgram-sdk
twilio
requests
httpx
